+++ packages/core/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @under-control/core
 
+## 1.7.12
+
+### Patch Changes
+
+- Improve type infer in `input` / `output` mappers
+
 ## 1.7.11
 
 ### Patch Changes