+++ packages/forms/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @under-control/forms
 
+## 1.7.12
+
+### Patch Changes
+
+- Improve type infer in `input` / `output` mappers
+- Updated dependencies
+  - @under-control/inputs@1.7.12
+  - @under-control/core@1.7.12
+  - @under-control/validate@1.7.12
+
 ## 1.7.11
 
 ### Patch Changes